% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{methods}
\name{dbWriteTable,GreenplumConnection,character,data.frame-method}
\alias{dbWriteTable,GreenplumConnection,character,data.frame-method}
\title{Greenplum dbWriteTable method}
\usage{
\S4method{dbWriteTable}{GreenplumConnection,character,data.frame}(conn, name,
  value, ..., row.names = FALSE, overwrite = FALSE, append = FALSE,
  field.types = NULL, temporary = FALSE, distributed_by = NULL,
  copy = TRUE)
}
\arguments{
\item{conn}{a [GreenplumConnection-class] object}

\item{name}{a character string specifying a table name. Names will be
automatically quoted so you can use any sequence of characters, not just
any valid bare table name.}

\item{value}{A data.frame to write to the database.}

\item{...}{Other arguments used by individual methods.}

\item{row.names}{Either TRUE, FALSE, NA or a string}

\item{overwrite}{a logical specifying whether to overwrite an existing table or not. Its default is FALSE.}

\item{append}{a logical specifying whether to append to an existing table in the DBMS. Its default is FALSE.}

\item{field.types}{character vector of named SQL field types where the names are the names of new table's columns.}

\item{temporary}{If TRUE, will generate a temporary table statement.}

\item{distributed_by}{Distribution columns for new table. NULL for random distribution.}

\item{copy}{If TRUE, data will be copied to remote database}
}
\description{
Greenplum dbWriteTable method
}
