% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreenplumDriver.R
\name{Greenplum}
\alias{Greenplum}
\title{Greenplum driver}
\usage{
Greenplum()
}
\description{
This driver never needs to be unloaded and hence `dbUnload()` is a
null-op.
}
\examples{
library(DBI)
RGreenplum::Greenplum()
}
