% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreenplumDriver.R
\docType{class}
\name{GreenplumDriver-class}
\alias{GreenplumDriver-class}
\alias{dbUnloadDriver,GreenplumDriver-method}
\title{GreenplumDriver and methods.}
\usage{
\S4method{dbUnloadDriver}{GreenplumDriver}(drv, ...)
}
\description{
GreenplumDriver and methods.
}
\keyword{internal}
