% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreenplumDriver.R
\docType{methods}
\name{dbIsValid,GreenplumDriver-method}
\alias{dbIsValid,GreenplumDriver-method}
\title{Greenplum dbIsValid method}
\usage{
\S4method{dbIsValid}{GreenplumDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{database object}

\item{...}{other arguments for method}
}
\description{
Greenplum dbIsValid method
}
