% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-create.R
\docType{methods}
\name{sqlCreateTable,GreenplumConnection-method}
\alias{sqlCreateTable,GreenplumConnection-method}
\title{Greenplum sqlCreateTable method}
\usage{
\S4method{sqlCreateTable}{GreenplumConnection}(con, table, fields,
  row.names = NA, temporary = FALSE, distributed_by = NULL, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{fields}{Either a character vector or a data frame.

A named character vector: Names are column names, values are types.
Names are escaped with \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.
Field types are unescaped.

A data frame: field types are generated using
\code{\link[=dbDataType]{dbDataType()}}.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{distributed_by}{Distribution columns for new table. NULL for random distribution.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Greenplum sqlCreateTable method
}
