% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{methods}
\name{dbWriteTable,GreenplumConnection,character,data.frame-method}
\alias{dbWriteTable,GreenplumConnection,character,data.frame-method}
\title{Greenplum dbWriteTable method}
\usage{
\S4method{dbWriteTable}{GreenplumConnection,character,data.frame}(conn, name,
  value, ..., row.names = FALSE, overwrite = FALSE, append = FALSE,
  field.types = NULL, temporary = FALSE, distributed_by = NULL,
  copy = TRUE)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{...}{Other parameters passed on to methods.}

\item{row.names}{Either TRUE, FALSE, NA or a string}

\item{overwrite}{a logical specifying whether to overwrite an existing table or not. Its default is FALSE.}

\item{append}{a logical specifying whether to append to an existing table in the DBMS. Its default is FALSE.}

\item{field.types}{character vector of named SQL field types where the names are the names of new table's columns.}

\item{temporary}{If TRUE, will generate a temporary table statement.}

\item{distributed_by}{Distribution columns for new table. NULL for random distribution.}

\item{copy}{If TRUE, data will be copied to remote database}
}
\description{
Greenplum dbWriteTable method
}
